configured the necessary access controls and security features

+++ app/Http/Controllers/PostsController.php
@@ -50,6 +50,7 @@ class PostsController extends Controller
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
         $post->save();
 
         return redirect('posts')->with('success', 'Post Created Successfully');
@@ -76,6 +77,11 @@ class PostsController extends Controller
      */
     public function edit(Post $post)
     {
+        if($post->user_id !== auth()->user()->id)
+        {
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -109,6 +115,11 @@ class PostsController extends Controller
      */
     public function destroy(Post $post)
     {
+        if($post->user_id !== auth()->user()->id)
+        {
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'Post deleted');
     }
